requirements.txt
# 物理实验数据采集软件依赖包
# 安装命令: pip install -r requirements.txt

PyQt6>=6.4.0
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0
